pyproject.toml
[project]
name = "radicle-natives"
version = "0.0.1"
description = "Automated integrations for operating Radicle Natives"
authors = [
    {name = "Anthony Shull"}
]
license = {text = "Free"}
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "questdb (>=3.0.0,<4.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "faker (>=37.11.0,<38.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "mkdocs (>=1.6.1,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "ruff (>=0.14.0,<0.15.0)"
]
